requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dash==3.1.1
distro==1.9.0
easyocr==1.7.2
filelock==3.18.0
Flask==3.1.1
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.48.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
openai==1.97.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.2.0
protobuf==6.31.1
pyarrow==21.0.0
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
retrying==1.4.1
rpds-py==0.26.0
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.47.0
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.6.11
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
zipp==3.23.0
